backend/requirements.txt
aniso8601==9.0.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JSON==0.4.0
Flask-JWT==0.3.2
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.28
packaging==23.2
protobuf==4.25.1
PyJWT==1.4.2
pytz==2023.3.post1
six==1.16.0
Werkzeug==3.0.1
